Fix formatting when there's an inline comment on the final line
This includes a test case to illustrate the issue.

The TLDR is that rewrite-clj treats comments as whitespace, so if there
is a trailing comment on the final line of a form, cljstyle would tell
you to remove the space between the form and the comment because it
considered the whole rest of the line to be whitespace.