Make ignore behavior match documentation
https://github.com/greglook/cljstyle/blob/bb79e3ad0a2db0541d137a81aedd40455de5c143/doc/configuration.md#file-settings

The configuration doc says this under the :ignore bullet point:

...patterns are matched against the entire (relative) file path...

However, this is not the case in the implementation, as
cljstyle.config/ignored? uses the canonical path, rather than a
relatived version of it.

To match the docs, this change relatives the canonical path to the file
against the JVM working dir. Hopefully this is the intended behavior :)