Make ignore behavior match documentation (#91)
The configuration doc says this under the :ignore bullet point:

...patterns are matched against the entire (relative) file path...

However, this is not the case in the implementation, as
cljstyle.config/ignored? uses the canonical path, rather than a
relatived version of it.

To match the docs, this change relatives the canonical path to the file
against the JVM working dir. Hopefully this is the intended behavior :)